Refactor: seperate handler & base response from common (#93)
