+++ src/main/java/com/gaduationproject/cre8/app/employmentpost/controller/EmployeePostSearchController.java
@@ -1,6 +1,6 @@
 package com.gaduationproject.cre8.app.employmentpost.controller;
 
-import com.gaduationproject.cre8.common.response.BaseResponse;
+import com.gaduationproject.cre8.app.response.BaseResponse;
 import com.gaduationproject.cre8.domain.employmentpost.search.EmployeePostSearch;
 import com.gaduationproject.cre8.app.employmentpost.dto.response.EmployeePostSearchWithCountResponseDto;
 import com.gaduationproject.cre8.app.employmentpost.service.EmployeePostSearchService;
