+++ src/test/java/com/gaduationproject/cre8/app/portfolio/controller/PortfolioControllerTest.java
@@ -1,16 +1,10 @@
 package com.gaduationproject.cre8.app.portfolio.controller;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doNothing;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 
-import com.gaduationproject.cre8.app.auth.config.SecurityConfig;
-import com.gaduationproject.cre8.app.auth.handler.JwtAccessDeniedHandler;
-import com.gaduationproject.cre8.app.auth.handler.JwtAuthenticationEntryPointHandler;
-import com.gaduationproject.cre8.app.auth.jwt.TokenProvider;
-import com.gaduationproject.cre8.app.member.controller.ProfileController;
 import com.gaduationproject.cre8.app.portfolio.dto.request.PortfolioEditRequestDto;
 import com.gaduationproject.cre8.app.portfolio.dto.response.PortfolioResponseDto;
 import com.gaduationproject.cre8.app.portfolio.dto.response.PortfolioSimpleResponseDto;
@@ -19,36 +13,24 @@ import com.gaduationproject.cre8.common.LocalDateDeserializer;
 import com.gaduationproject.cre8.common.LocalDateSerializer;
 import com.gaduationproject.cre8.common.TestSecurityConfig;
 import com.gaduationproject.cre8.common.WithMockCustomUser;
-import com.gaduationproject.cre8.common.response.error.RestExceptionHandler;
 import com.gaduationproject.cre8.domain.member.entity.Member;
 import com.gaduationproject.cre8.domain.member.type.Sex;
 import com.gaduationproject.cre8.domain.portfolio.entity.Portfolio;
 import com.gaduationproject.cre8.domain.portfolio.entity.PortfolioImage;
 import com.gaduationproject.cre8.domain.workfieldtag.entity.WorkFieldTag;
-import com.gaduationproject.cre8.externalApi.redis.service.RedisUtil;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import java.awt.image.ImagingOpException;
-import java.io.IOException;
 import java.time.LocalDate;
 import java.util.List;
-import org.hibernate.jdbc.Work;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.method.annotation.AuthenticationPrincipalArgumentResolver;
 import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.MockMvcBuilder;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
